Fixed dependency rules ignoring blacklist.
Now if a package or component is blacklisted it will not be built even if
something else is marked as depending on it. Also, explicity specifying the
packages and components when invoking make will not now build the targets they
depend on. In summary package dependencies now respect the blacklist and the
specified set of packages to be built.