Created generic FIFO to be used for messaging.
The FIFO queue is split into 'put' and 'get' ends. Each end has a separate
owner and can exist independently of the other. When one end is destroyed the
FIFO is automatically broken. The FIFO queue is a more generic replacement for
the unfinished mailbox implementation that was previously to be used for
messaging. As such, the mailbox source has been removed with this commit.