First pass at message passing system.
Fixed const correctness for AutoReadLock. Fixed bug with FIFO slicing objects
put into it when it makes a copy. Also updated message generation script to add
a clone method to every method so they can be passed through the fixed FIFO
pipes. Finally, the message generation script now adds an enum of the message
types and a matches function to each message to check whether it matches the
specified message types.