Fixed serious bug in spinlock.
Previously, if the CMPXCHG instruction did not acquire the lock first time (ie
it was already held) it would set the AL register to the value of the non-zero
value of the lock. The next instruction compared AL to zero to test whether the
lock had been acquired. In this case it would find AL to be non-zero and jump
to the top of the spinlock. On this and all following iterations where the lock
remains held (has non-zero value) the CMPXCHG would compare AL to the lock
value, see that they match and set the lock value to non-zero (essentially do
nothing). The problem arises when whatever was holding the lock finally
releases it. On the following iteration CMPXCHG compares AL (non-zero) to the
lock value (now zero), sees that they don't match and sets AL to the lock
value. Now AL is zero and the spinlock can exit but the lock value also remains
zero. So the lock wasn't really acquired. This bug is a result of not setting
AL to zero at the beginning of every iteration (just the first).