Fixed crash bug in physics simulation.
When the player attempted to rotate both ways at once, equal and opposite spins
would both be applied. These would cancel each other out and result in the spin
quaternion being the identity. However the "spin applied" flag would still be
set causing the physics simulation to try to integrate the rotational motion. It
would extract the zero vector as the axis of rotation. When it normalised this
NaN's were introduced into the system which later caused Ogre3D to crash when
rendering. The fix is two fold. First the rotational motion integration function
checks it is not trying to normalise a zero vector and second the control input
system will avoid applying any spin if the player is trying to turn both ways at
once.