Fixed KDTree construction bug.
Previously the split less-than functor determined split A to be less than split
B if its position was less than B's position by at least a small tolerence
value. However, if it was within this tolerence range its position was judged
to be the same as B's and the less-than functor deferred ordering to the split
axes (X < Y < Z). This implementation gave rise to improperly sorted split
planes when building a KDTree from triangles with distinct vertices that were
close enough together along one axis to fall within the tolerence range of the
old less-than functor.  The improperly sorted split planes caused assumptions
in the triangle counting code to be violated and an assert raised. This fix
simply removes the tolerance range and requires split A's position to be
strictly equal to split B's before it will defer to ordering by split axis.