Fixed bug with initial node cost in KDTree.
When there are many triangles the split cost is higher than the arbitrary cost
of 100 chosen for the initial node. Thus the construction algorithm would deem
it cheaper not to split the initial node at all.  This was fixed by setting the
cost of the initial node to the max float value possible.