Fixed particle system not being destroyed.
Still not quite sure what caused this bug. Apparently the temporary auto_ptr in
the foreach loop was null when it went out of scope so nothing was deleted.
Giving the auto_ptr an explicit name solved the problem. Further investigation
should be done at some point because the temporary auto_ptr method should work.