Refactored physics module to support game objects.
The new class hierarchy has missiles, ships and (eventually) other objects
deriving from the GameObject class. This in turn derives from the RigidBody
class which handles the physics simulation. The GameObject class provides some
additional functionality not strictly to do with physics simulation and thus
not found in RigidBody. This includes double-dispatch boilerplate code to
handle collisions between various object types. Note that this commit is
a work-in-progress. While the refactoring of the physics module is
provisionally complete (it compiles at least), the rest of the codebase has
yet to be updated and is currently broken to some degree as a result.